requirements.txt
aiogram==2.25.1
asyncpg==0.24.0
fastapi==0.91.0
pydantic==1.10.4
python-dotenv==0.21.1
requests==2.27.1
uvicorn>=0.15.0,<0.16.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dateutil==2.8.2
pandas==1.3.5
openpyxl==3.1.2
sqlmodel==0.0.4
python-multipart==0.0.6
Pillow==9.4.0
numpy
opencv-python==4.5.2.52
matplotlib==3.1.3
tensorflow==2.12.0
torch==2.0.1
torchvision==0.15.2
easyocr==1.7.0